Add delete task function
